data-engine/requirements.txt
fastapi
uvicorn
python-dotenv
requests
yfinance
pandas
lxml
scipy
